split out MatPES set generators and jobs from mp.py into separate modules
